client_credit_union/report_generator/pyproject.toml
[project]
name = "report-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "docxtpl>=0.20.0",
    "matplotlib>=3.10.3",
    "numpy>=2.2.6",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pyyaml>=6.0.2",
    "seaborn>=0.13.2",
    "typer>=0.16.0",
]

[dependency-groups]
dev = [
    "ruff>=0.11.13",
]
